requirements.txt
streamlit
pydantic
scikit-learn
transformers
torch
python-dotenv